Rename eu.xenit.docker-compose plugins
Use eu.xenit.docker-compose as a base plugin and eu.xenit.docker-compose.auto as an automatic configuration plugin.
This naming should make it more clear to the user that a lot of automatic configuration is going on,
like adding all DockerBuildImage tasks as dependency of the ComposeUp task when using the automatic configuration plugin.
