+++ src/integrationTest/examples/docker-compose-auto-example/settings.gradle
