+++ src/integrationTest/examples/docker-compose-auto-example/subpr1/build.gradle
