+++ src/integrationTest/examples/docker-compose-auto-example/subpr2/build.gradle
