--- src/integrationTest/examples/docker-compose-base-example/build.gradle
@@ -1,14 +1,0 @@
-plugins {
-    id "eu.xenit.docker-compose.base"
-}
-evaluationDependsOn(":subpr2")
-
-dockerCompose {
-    fromProject(':subpr1')
-    fromBuildImage("TEST_IMAGE", project.project(':subpr2').tasks.named("buildDockerImage"))
-}
-
-task integrationTest {
-}
-
-dockerCompose.isRequiredBy(integrationTest)
