+++ src/integrationTest/examples/docker-compose-example/build.gradle
@@ -1,6 +1,12 @@
 plugins {
     id "eu.xenit.docker-compose"
 }
+evaluationDependsOn(":subpr2")
+
+dockerCompose {
+    fromProject(':subpr1')
+    fromBuildImage("TEST_IMAGE", project.project(':subpr2').tasks.named("buildDockerImage"))
+}
 
 task integrationTest {
 }
