+++ src/main/java/eu/xenit/gradle/docker/compose/DockerComposeAutoPlugin.java
@@ -1,0 +1,18 @@
+package eu.xenit.gradle.docker.compose;
+
+import org.gradle.api.Plugin;
+import org.gradle.api.Project;
+
+public class DockerComposeAutoPlugin implements Plugin<Project> {
+
+    @Override
+    public void apply(Project project) {
+        project.getPlugins().withType(DockerComposePlugin.class, dockerComposePlugin -> {
+            project.allprojects(project1 -> {
+                dockerComposePlugin.getDockerComposeConvention().fromProject(project1);
+            });
+        });
+
+        project.getPlugins().apply(DockerComposePlugin.class);
+    }
+}
